+++ src/helpers/ls.ts
@@ -2,12 +2,12 @@ export const setItem = (name: string, value: any): void => {
   if (Array.isArray(value) || (typeof value === 'object' && value !== null)) {
     value = JSON.stringify(value) as string
   }
-  localStorage.setItem(name, value as string)
+  localStorage.setItem(name, String(value))
 }
 
 export const getItem = (name: string): any => {
   const value = localStorage.getItem(name)
-  if (!value) return null
+  if (value == null) return null
   if (['true', 'false'].includes(value)) return value === 'true'
   if (!isNaN(Number(value))) return Number(value)
   try {