+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved scripts and tests processing
+
 ## [1.0.0-beta.3] - 2023-07-22
 
 ### Changed